python_server/pyproject.toml
[project]
name = "python_server"
version = "0.1.0"
description = "Optimized RAG Pipeline with Flask"
authors = [
    { name = "Jishnu Duhan", email = "duhanjishnuemail@gmail.com" },
]
dependencies = [
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.5.0",
    "pypdf2>=3.0.1",
    "langchain>=0.0.346",
    "langchain-community>=0.0.14",
    "faiss-cpu>=1.7.4",
    "langchain-ollama>=0.1.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "upstash-vector>=0.8.0",
]
requires-python = ">=3.11"
